requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.3.0
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.7
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.16
environ==1.0
et-xmlfile==1.1.0
executing==2.1.0
Faker==18.11.2
idna==3.7
ipykernel==6.29.5
ipython==8.27.0
isort==5.12.0
jedi==0.19.1
jmespath==1.0.1
JPype1==1.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
Markdown==3.6
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==2.1.1
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
platformdirs==4.2.2
prompt-toolkit==3.0.41
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.0
pypdf==5.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.31.0
rsa==4.7.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tabula==1.0.5
tabula-py==2.9.3
tabulate==0.9.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
